backend/requirements/base.txt
Django==5.0.14
psycopg2-binary==2.9.11
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
django-environ==0.11.2
django-encrypted-model-fields==0.6.5
django-phonenumber-field[phonenumbers]==7.2.0
pypdf>=3.9.0
pydantic==2.5.3
cryptography==44.0.1
